feat(#84):[Tilak|Pooja] use seperate modules for wallet and verifier and remove online code
